新增多种序列化器以及SPI机制

+++ example-common/src/main/java/com/cgxdgfcd/example/common/service/UserService.java
@@ -20,7 +20,7 @@ public interface UserService {
      *
      * @return
      */
-    default short getNumber() {
+    default int getNumber() {
         return 1;
     }
 }
